Makefile cleanups: Make clean now removes all object code and generated source
Driver tested up to and including M3 generation.
  - Bug in Calculus: Somehow, the atom A=A was sneaking into the compilation process and messing up the variable substitution algorithm; Self-equality atoms are now filtered out beforehand.
  - Bug in CalcToM3: LHS variables in the trigger were being taken from the map declaration, not the output of compile_delta_for_rel.  Still need to fix the corresponding bits in the map initialization code

Three known bugs still remain in the driver
  - Need to get the initialization code to use variable names as defined by the map declaration
  - The code currently produces special maps representing each relation (for initialization), but these maps are created even if we can ignore the initialization code for a given map (ie, there are no input variables).
  - We currently compute variable mappings for substituted map accesses by comparing the bindings in the trigger schemas.  However, the code that decides whether two bits of code are equivalent also returns a set of substitutions.  These two elements should be used as a sanity check on each other.
  - Not a bug, but we're currently producing a whole mess of debug output that needs to go away.



git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@861 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 220a6923f95b1ef294b8a18bb60da27a7d3bf891

+++ mmdb/dbtoaster/prototype/compiler/alpha4/test/rst.sql
@@ -1,9 +1,9 @@
 
-CREATE TABLE R(A double, B double) 
+CREATE TABLE R(A int, B int) 
   FROM FILE 'test/data/r.dat' LINE DELIMITED CSV;
-CREATE TABLE S(B double, C double) 
+CREATE TABLE S(B int, C int) 
   FROM FILE 'test/data/s.dat' LINE DELIMITED CSV;
-CREATE TABLE T(C double, D double)
+CREATE TABLE T(C int, D int)
   FROM FILE 'test/data/t.dat' LINE DELIMITED CSV;
 
 SELECT sum(A*D) FROM R,S,T WHERE R.B=S.B AND S.C=T.C;
